\name{scapeMCMC-package}
\alias{scapeMCMC-package}
\alias{scapeMCMC}
\docType{package}
\title{MCMC Diagnostic Plots}
\description{
  Markov-chain Monte Carlo diagnostic plots, accompanying the
  \pkg{scape} package. The purpose of the package is to combine existing
  tools from the \pkg{coda} and \pkg{lattice} packages, and make it easy
  to adjust graphical details. It can be useful for anyone using MCMC
  analysis, regardless of the application.
}
\details{
  \code{help(package="scapeMCMC")} gives an overview of the package.

  The \pkg{scape} vignette \file{scape/doc/dsc.pdf} provides background
  and introduces the package design, but is not updated.
}
\author{Arni Magnusson \email{arnima@u.washington.edu} and Ian Stewart.}
% Graphics
\keyword{hplot}
% Statistics
\keyword{ts}
